♻️  -- Submission view ->ViewSet. Fixed tests. Removed instance property and used * reference.
